Sendding the methods to externald files.
I was send the methods in the shopping screen to other files.
